feat: fix compiling error for MultiplayerDebugAuditTrail

+++ Code/Framework/AzFramework/AzFramework/Quality/QualityCVarGroup.cpp
@@ -214,9 +214,9 @@ namespace AzFramework
                 }
                 else if (!result)
                 {
-                    // the requested qualityLevel index wasn't found, use highest available 
+                    // the requested qualityLevel index wasn't found, use highest available
                     PerformConsoleCommand(command, FixedValueString::format("%.*s/%d",
-                        AZ_STRING_ARG(visitArgs.m_jsonKeyPath), currentQualityLevel));
+                        AZ_STRING_ARG(visitArgs.m_jsonKeyPath), static_cast<int>(currentQualityLevel)));
                 }
             }
             else
